+++ tests/testthat/test_wa_data.R
@@ -18,12 +18,35 @@ test_that("wa_locked_out", {
   expect_gt(terra::global(wa_locked_out, "max", na.rm = TRUE)[[1]], 0)
 })
 
-test_that("wa_features", {
-  wa_features <- get_wa_features()
-  expect_is(wa_features, "SpatRaster")
+test_that("wa_species", {
+  wa_species <- get_wa_species()
+  expect_is(wa_species, "SpatRaster")
   expect_true(
-    all(terra::global(wa_features, "min", na.rm = TRUE)[[1]] >= 0)
+    all(terra::global(wa_species, "min", na.rm = TRUE)[[1]] >= 0)
   )
+})
+
+test_that("wa_carbon", {
+  wa_carbon <- get_wa_carbon()
+  expect_is(wa_carbon, "SpatRaster")
+  expect_equal(terra::nlyr(wa_carbon), 1)
+  expect_true(
+    all(terra::global(wa_carbon, "min", na.rm = TRUE)[[1]] >= 0)
+  )
+})
+
+test_that("wa_attr", {
+  wa_attr <- get_wa_attr()
+  expect_is(wa_attr, "tbl_df")
+  expect_named(
+    wa_attr,
+    c(
+      "feature", "binomial", "family", "order",
+      "extinction_prob", "interest_score"
+    )
+  )
+  expect_equal(wa_attr$feature, names(get_wa_species()))
+  expect_equal(nrow(wa_attr), nrow(na.omit(wa_attr)))
 })
 
 test_that("locked data are feasible", {
@@ -52,17 +75,20 @@ test_that("standardized missing values", {
   wa_pu <- get_wa_pu()
   wa_locked_in <- get_wa_locked_in()
   wa_locked_out <- get_wa_locked_out()
-  wa_features <- get_wa_features()
+  wa_species <- get_wa_species()
+  wa_carbon <- get_wa_carbon()
   expect_equal(nonNA_cells(wa_pu), nonNA_cells(wa_locked_in))
   expect_equal(nonNA_cells(wa_pu), nonNA_cells(wa_locked_out))
-  expect_equal(nonNA_cells(wa_pu), nonNA_cells(sum(wa_features)))
+  expect_equal(nonNA_cells(wa_pu), nonNA_cells(sum(wa_species)))
+  expect_equal(nonNA_cells(wa_pu), nonNA_cells(wa_carbon))
 })
 
 test_that("rasters are comparable", {
   wa_pu <- get_wa_pu()
   wa_locked_in <- get_wa_locked_in()
   wa_locked_out <- get_wa_locked_out()
-  wa_features <- get_wa_features()
+  wa_species <- get_wa_species()
+  wa_carbon <- get_wa_carbon()
   expect_true(
     terra::compareGeom(wa_pu, wa_locked_out, stopOnError = FALSE)
   )
@@ -70,6 +96,9 @@ test_that("rasters are comparable", {
     terra::compareGeom(wa_pu, wa_locked_in, stopOnError = FALSE)
   )
   expect_true(
-    terra::compareGeom(wa_pu, wa_features, stopOnError = FALSE)
+    terra::compareGeom(wa_pu, wa_species, stopOnError = FALSE)
+  )
+  expect_true(
+    terra::compareGeom(wa_pu, wa_carbon, stopOnError = FALSE)
   )
 })